requirements.txt
Django>=3.0.8,<3.1.0
djangorestframework>=3.11.0,<3.12.0
psycopg2>=2.8.5,<2.9.0

Pillow>=7.2.0,<7.3.0

schedule>=0.6.0,<0.7.0

celery>=4.4.7,<4.5.0

django-celery-results>=1.2.1,<1.3.0
django-celery-beat>=2.0.0,<2.1.0

redis>=3.5.3,<3.6.0
#kombu>=5.0.0,<5.1.0
pyPdf>=1.13,<1.14
reportlab>=3.5.48,<3.6.0
python-crontab>=2.5.1,<2.6.0

django-widget-tweaks>=1.4.8,<1.5.0

xhtml2pdf>=0.2.4,<0.3.0
